+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/network/item/NetworkItemWirelessCraftingMonitor.java
@@ -5,13 +5,11 @@ import com.raoulvdberge.refinedstorage.RSGui;
 import com.raoulvdberge.refinedstorage.api.network.INetwork;
 import com.raoulvdberge.refinedstorage.api.network.item.INetworkItem;
 import com.raoulvdberge.refinedstorage.api.network.item.INetworkItemHandler;
-import com.raoulvdberge.refinedstorage.api.network.item.NetworkItemAction;
 import com.raoulvdberge.refinedstorage.api.network.security.Permission;
 import com.raoulvdberge.refinedstorage.item.ItemWirelessCraftingMonitor;
 import com.raoulvdberge.refinedstorage.util.WorldUtils;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.EnumHand;
 import net.minecraftforge.energy.CapabilityEnergy;
 import net.minecraftforge.energy.IEnergyStorage;
 
@@ -19,11 +17,13 @@ public class NetworkItemWirelessCraftingMonitor implements INetworkItem {
     private INetworkItemHandler handler;
     private EntityPlayer player;
     private ItemStack stack;
+    private int invIndex;
 
-    public NetworkItemWirelessCraftingMonitor(INetworkItemHandler handler, EntityPlayer player, ItemStack stack) {
+    public NetworkItemWirelessCraftingMonitor(INetworkItemHandler handler, EntityPlayer player, ItemStack stack, int invIndex) {
         this.handler = handler;
         this.player = player;
         this.stack = stack;
+        this.invIndex = invIndex;
     }
 
     @Override
@@ -32,7 +32,7 @@ public class NetworkItemWirelessCraftingMonitor implements INetworkItem {
     }
 
     @Override
-    public boolean onOpen(INetwork network, EntityPlayer player, EnumHand hand) {
+    public boolean onOpen(INetwork network) {
         if (RS.INSTANCE.config.wirelessCraftingMonitorUsesEnergy && stack.getItemDamage() != ItemWirelessCraftingMonitor.TYPE_CREATIVE && stack.getCapability(CapabilityEnergy.ENERGY, null).getEnergyStored() <= RS.INSTANCE.config.wirelessCraftingMonitorOpenUsage) {
             return false;
         }
@@ -43,7 +43,7 @@ public class NetworkItemWirelessCraftingMonitor implements INetworkItem {
             return false;
         }
 
-        player.openGui(RS.INSTANCE, RSGui.WIRELESS_CRAFTING_MONITOR, player.getEntityWorld(), hand.ordinal(), network.world().provider.getDimension(), 0);
+        player.openGui(RS.INSTANCE, RSGui.WIRELESS_CRAFTING_MONITOR, player.getEntityWorld(), invIndex, 0, 0);
 
         drainEnergy(RS.INSTANCE.config.wirelessCraftingMonitorOpenUsage);
 
@@ -51,25 +51,14 @@ public class NetworkItemWirelessCraftingMonitor implements INetworkItem {
     }
 
     @Override
-    public void onAction(NetworkItemAction action) {
-        switch (action) {
-            case CRAFTING_TASK_CANCELLED:
-                drainEnergy(RS.INSTANCE.config.wirelessCraftingMonitorCancelUsage);
-                break;
-            case CRAFTING_TASK_ALL_CANCELLED:
-                drainEnergy(RS.INSTANCE.config.wirelessCraftingMonitorCancelAllUsage);
-                break;
-        }
-    }
-
-    private void drainEnergy(int energy) {
+    public void drainEnergy(int energy) {
         if (RS.INSTANCE.config.wirelessCraftingMonitorUsesEnergy && stack.getItemDamage() != ItemWirelessCraftingMonitor.TYPE_CREATIVE) {
             IEnergyStorage energyStorage = stack.getCapability(CapabilityEnergy.ENERGY, null);
 
             energyStorage.extractEnergy(energy, false);
 
             if (energyStorage.getEnergyStored() <= 0) {
-                handler.onClose(player);
+                handler.close(player);
 
                 player.closeScreen();
             }