+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/network/item/NetworkItemWirelessGrid.java
@@ -4,15 +4,14 @@ import com.raoulvdberge.refinedstorage.RS;
 import com.raoulvdberge.refinedstorage.api.network.INetwork;
 import com.raoulvdberge.refinedstorage.api.network.item.INetworkItem;
 import com.raoulvdberge.refinedstorage.api.network.item.INetworkItemHandler;
-import com.raoulvdberge.refinedstorage.api.network.item.NetworkItemAction;
 import com.raoulvdberge.refinedstorage.api.network.security.Permission;
 import com.raoulvdberge.refinedstorage.apiimpl.API;
 import com.raoulvdberge.refinedstorage.item.ItemWirelessGrid;
 import com.raoulvdberge.refinedstorage.tile.grid.WirelessGrid;
 import com.raoulvdberge.refinedstorage.util.WorldUtils;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.EnumHand;
 import net.minecraftforge.energy.CapabilityEnergy;
 import net.minecraftforge.energy.IEnergyStorage;
 
@@ -33,7 +32,7 @@ public class NetworkItemWirelessGrid implements INetworkItem {
     }
 
     @Override
-    public boolean onOpen(INetwork network, EntityPlayer player, EnumHand hand) {
+    public boolean onOpen(INetwork network) {
         if (RS.INSTANCE.config.wirelessGridUsesEnergy && stack.getItemDamage() != ItemWirelessGrid.TYPE_CREATIVE && stack.getCapability(CapabilityEnergy.ENERGY, null).getEnergyStored() <= RS.INSTANCE.config.wirelessGridOpenUsage) {
             return false;
         }
@@ -44,7 +43,7 @@ public class NetworkItemWirelessGrid implements INetworkItem {
             return false;
         }
 
-        API.instance().openWirelessGrid(player, hand, network.world().provider.getDimension(), WirelessGrid.ID);
+        API.instance().getGridManager().openGrid(WirelessGrid.ID, (EntityPlayerMP) player, stack);
 
         drainEnergy(RS.INSTANCE.config.wirelessGridOpenUsage);
 
@@ -52,25 +51,14 @@ public class NetworkItemWirelessGrid implements INetworkItem {
     }
 
     @Override
-    public void onAction(NetworkItemAction action) {
-        switch (action) {
-            case ITEM_INSERTED:
-                drainEnergy(RS.INSTANCE.config.wirelessGridInsertUsage);
-                break;
-            case ITEM_EXTRACTED:
-                drainEnergy(RS.INSTANCE.config.wirelessGridExtractUsage);
-                break;
-        }
-    }
-
-    private void drainEnergy(int energy) {
+    public void drainEnergy(int energy) {
         if (RS.INSTANCE.config.wirelessGridUsesEnergy && stack.getItemDamage() != ItemWirelessGrid.TYPE_CREATIVE) {
             IEnergyStorage energyStorage = stack.getCapability(CapabilityEnergy.ENERGY, null);
 
             energyStorage.extractEnergy(energy, false);
 
             if (energyStorage.getEnergyStored() <= 0) {
-                handler.onClose(player);
+                handler.close(player);
 
                 player.closeScreen();
             }