+++ src/main/java/com/raoulvdberge/refinedstorage/proxy/ProxyClient.java
@@ -2,6 +2,7 @@ package com.raoulvdberge.refinedstorage.proxy;
 
 import com.raoulvdberge.refinedstorage.RSKeyBindings;
 import com.raoulvdberge.refinedstorage.block.BlockBase;
+import com.raoulvdberge.refinedstorage.gui.KeyInputListener;
 import com.raoulvdberge.refinedstorage.item.ItemBase;
 import com.raoulvdberge.refinedstorage.render.IModelRegistration;
 import com.raoulvdberge.refinedstorage.render.collision.BlockHighlightListener;
@@ -45,6 +46,7 @@ public class ProxyClient extends ProxyCommon implements IModelRegistration {
         super.preInit(e);
 
         MinecraftForge.EVENT_BUS.register(new BlockHighlightListener());
+        MinecraftForge.EVENT_BUS.register(new KeyInputListener());
     }
 
     @Override